test(tests): split model-matching-core by behaviour family
Split model-matching-core.test.js (419 LOC) into
model-matching-score.test.js (calculateJobMatch) and
model-matching-discovery.test.js (findMatchingJobs,
deriveDevelopmentPath, findNextStepJob) along the inner describe
boundaries. Shared deriveJob fixture reproduced per sibling. Assertions
unchanged; 17 tests preserved.

model-validation-data.test.js is retained over the ceiling on the
allow-list — its tests all exercise the single validateAllData contract
through one entry point with no existing describe seam.

Spec 0640 § C (large files), design Decision 5.
